# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Emerj Pro API",
        "description": "API Documentation of Emerj Pro System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "description": "User registration data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "dob",
                                    "ibi_id",
                                    "ibi_name",
                                    "phone_number",
                                    "address_line1",
                                    "city",
                                    "state",
                                    "country",
                                    "region",
                                    "subscription_type_id",
                                    "currency",
                                    "price",
                                    "start_date",
                                    "role_id"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "Username",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "first_name": {
                                        "description": "First name",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "last_name": {
                                        "description": "Last name",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "dob": {
                                        "description": "Date of birth (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "ibi_id": {
                                        "description": "IBI ID",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "ibi_name": {
                                        "description": "IBI name",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "phone_number": {
                                        "description": "Phone number",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "notes": {
                                        "description": "Notes",
                                        "type": "string"
                                    },
                                    "register_at": {
                                        "description": "Registration date (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "profile": {
                                        "description": "Profile image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "address_line1": {
                                        "description": "Address line 1",
                                        "type": "string"
                                    },
                                    "address_line2": {
                                        "description": "Address line 2",
                                        "type": "string"
                                    },
                                    "city": {
                                        "description": "City",
                                        "type": "string"
                                    },
                                    "state": {
                                        "description": "State",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string"
                                    },
                                    "region": {
                                        "description": "Region",
                                        "type": "string"
                                    },
                                    "subscription_type_id": {
                                        "description": "Subscription type ID",
                                        "type": "integer"
                                    },
                                    "currency": {
                                        "description": "Currency",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "start_date": {
                                        "description": "Start date (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "role_id": {
                                        "description": "Role ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "description": "Bearer token",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "User login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "token": {
                                                    "description": "Bearer token",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Login failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Regresh Token",
                "operationId": "f6409cd8d794404a6410b0767432b0f4",
                "responses": {
                    "200": {
                        "description": "New Token Generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not refresh token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt_auth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}